Add manage_service parameter to all agents::ml2 classes that did not yet have it.
This existed in the older agents::ovs and agents::linuxbridge classes
which were recently removed, adding here for consistency.

When set to false, enables puppet to configure a service without
starting/stopping it on each run.  This may be necessary when
using an external clustering system (Corosync/Pacemaker, for
example).  Defaults to true.

Change-Id: I3242f790e7d3231a81b447c10aa2d741b7b68a3e